[orchagent]: Separate global SAI API pointers and initialization function
- To lighten the weight of main.cpp and make it easier for future refactor
