+++ orchagent/saihelper.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void initSaiApi();
